+++ CHANGELOG.md
@@ -6,12 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.15.1] - 2023-07-29
+### Added
+- [@mjkuranda](https://github.com/mjkuranda): Space char in `InputMenuOption`.
+
 ### Changed
 - [@mjkuranda](https://github.com/mjkuranda): Spawning entity coordinates.
 
 ### Fixed
 - [@mjkuranda](https://github.com/mjkuranda): Player collision with entities.
-- 
 
 ## [0.15.0] - 2023-07-28
 ### Added